Move the thread fetch down to the Workspace component, add a callback to refresh the thread list when a new thread is created
